split out parsing
